requirements.txt
streamlit>=1.28.0
assemblyai>=0.21.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
pathlib2>=2.3.7
python-dotenv>=1.0.0
